# api/openapi.yaml
openapi: 3.1.0
info:
  title: Worthly API
  version: 4.3.0
  description: |
    Canonical MVP API contract for Worthly.

    Protected resources are Bearer-token OAuth2 Resource Server endpoints.
    The Next.js `web_session` cookie is never an API credential.

    Aggregate amounts are always grouped by ISO 4217 currency. No schema
    exposes a single cross-currency Money total.
servers:
  - url: https://worthly.example.com/api/v1
security:
  - oauth2: [worthly.read]
...
paths:
...
  /categories:
    get:
      tags: [Categories]
      operationId: listCategories
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Category'}}
    post:
      tags: [Categories]
      operationId: createCategory
      security: [{oauth2: [worthly.write]}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CategoryCreate'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Category'}
  /categories/{categoryId}:
    patch:
      tags: [Categories]
      operationId: updateCategory
      security: [{oauth2: [worthly.write]}]
      parameters:
        - {$ref: '#/components/parameters/CategoryId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CategoryPatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Category'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Categories]
      operationId: deleteCategory
      security: [{oauth2: [worthly.write]}]
      parameters:
        - {$ref: '#/components/parameters/CategoryId'}
      responses:
        '204': {description: Deleted}
        '404': {$ref: '#/components/responses/NotFound'}
  /categorization-rules:
    get:
      tags: [Categories]
      operationId: listRules
      responses:
        '200':
          description: Rules
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/CategorizationRule'}}
    post:
      tags: [Categories]
      operationId: createRule
      security: [{oauth2: [worthly.write]}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CategorizationRuleCreate'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CategorizationRule'}
  /categorization-rules/{ruleId}:
    patch:
      tags: [Categories]
      operationId: updateRule
      security: [{oauth2: [worthly.write]}]
      parameters:
        - {$ref: '#/components/parameters/RuleId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CategorizationRulePatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CategorizationRule'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Categories]
      operationId: deleteRule
      security: [{oauth2: [worthly.write]}]
      parameters:
        - {$ref: '#/components/parameters/RuleId'}
      responses:
        '204': {description: Deleted}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://worthly.example.com/oauth2/authorize
          tokenUrl: https://worthly.example.com/oauth2/token
          refreshUrl: https://worthly.example.com/oauth2/token
          scopes:
            openid: OpenID identity
            profile: Basic owner profile
            worthly.read: Read Worthly financial data
            worthly.write: Modify Worthly metadata/settings
  parameters:
...
    RuleId: {name: ruleId, in: path, required: true, schema: {type: string, format: uuid}}
    CategoryId: {name: categoryId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
...
    NotFound:
      description: Not found
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
...
  schemas:
...
    Category:
      type: object
      required: [id, label, system]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        label: {type: string}
        parentId: {type: [string, 'null'], format: uuid}
        system: {type: boolean}
    CategoryCreate:
      type: object
      required: [label]
      properties:
        label: {type: string, minLength: 1, maxLength: 80}
        parentId: {type: [string, 'null'], format: uuid}
    CategoryPatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        label: {type: string, minLength: 1, maxLength: 80}
        parentId: {type: string, format: uuid}
    RuleField:
      type: string
      enum: [MERCHANT, DESCRIPTION, ACCOUNT_ID, DIRECTION]
    RuleOperator:
      type: string
      enum: [EQUALS, CONTAINS, STARTS_WITH]
    CategorizationRule:
      type: object
      required: [id, priority, field, operator, matchValue, targetCategoryId, enabled]
      properties:
        id: {type: string, format: uuid}
        priority: {type: integer, minimum: 0}
        field: {$ref: '#/components/schemas/RuleField'}
        operator: {$ref: '#/components/schemas/RuleOperator'}
        matchValue: {type: string, minLength: 1, maxLength: 200}
        amountMin: {type: [string, 'null']}
        amountMax: {type: [string, 'null']}
        targetCategoryId: {type: string, format: uuid}
        enabled: {type: boolean}
    CategorizationRuleCreate:
      type: object
      required: [priority, field, operator, matchValue, targetCategoryId]
      additionalProperties: false
      properties:
        priority: {type: integer, minimum: 0}
        field: {$ref: '#/components/schemas/RuleField'}
        operator: {$ref: '#/components/schemas/RuleOperator'}
        matchValue: {type: string, minLength: 1, maxLength: 200}
        amountMin: {type: [string, 'null']}
        amountMax: {type: [string, 'null']}
        targetCategoryId: {type: string, format: uuid}
        enabled: {type: boolean, default: true}
    CategorizationRulePatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        priority: {type: integer, minimum: 0}
        field: {$ref: '#/components/schemas/RuleField'}
        operator: {$ref: '#/components/schemas/RuleOperator'}
        matchValue: {type: string, minLength: 1, maxLength: 200}
        amountMin: {type: [string, 'null']}
        amountMax: {type: [string, 'null']}
        targetCategoryId: {type: string, format: uuid}
        enabled: {type: boolean}
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        correlationId: {type: string}